Move index.js and main into src/
