+++ src/main/autoUpdate.js
