+++ src/main/browser.js
