+++ src/main/settings.js
